requirements.txt
aniso8601==2.0.0
click==6.7
Flask==0.12.2
Flask-JWT-Extended==3.6.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
passlib==1.7.1
PyJWT==1.5.3
PyMySQL==0.8.0
pytz==2017.3
six==1.11.0
SQLAlchemy==1.2.2
Werkzeug==0.14.1